Move layout-related code into text_layout.rs
